editor commands - move API commands to workbench core
