+++ src/vs/workbench/contrib/notebook/browser/diff/notebookDiffActions.ts
@@ -7,8 +7,7 @@ import { IBulkEditService, ResourceTextEdit } from 'vs/editor/browser/services/b
 import { localize } from 'vs/nls';
 import { Action2, MenuId, registerAction2 } from 'vs/platform/actions/common/actions';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
-import { viewColumnToEditorGroup } from 'vs/workbench/api/common/shared/editor';
-import { ActiveEditorContext } from 'vs/workbench/common/editor';
+import { ActiveEditorContext, viewColumnToEditorGroup } from 'vs/workbench/common/editor';
 import { CellDiffViewModel } from 'vs/workbench/contrib/notebook/browser/diff/celllDiffViewModel';
 import { NotebookTextDiffEditor } from 'vs/workbench/contrib/notebook/browser/diff/notebookTextDiffEditor';
 import { NotebookDiffEditorInput } from 'vs/workbench/contrib/notebook/browser/notebookDiffEditorInput';
